requirements.txt
pytest
attrs==19.3.0
aws-cdk.assets==1.28.0
aws-cdk.aws-apigateway==1.28.0
aws-cdk.aws-cloudwatch==1.28.0
aws-cdk.aws-elasticsearch==1.28.0
aws-cdk.aws-elasticache==1.28.0
aws-cdk.aws-events==1.28.0
aws-cdk.aws-iam==1.28.0
aws-cdk.aws-ec2==1.28.0
aws-cdk.aws-kms==1.28.0
aws-cdk.aws-lambda==1.28.0
aws-cdk.aws-logs==1.28.0
aws-cdk.aws-s3-assets==1.28.0
aws-cdk.aws-s3==1.28.0
aws-cdk.aws-ssm==1.28.0
aws-cdk.core==1.28.0
aws-cdk.cx-api==1.28.0
aws-cdk.region-info==1.28.0
boto3==1.12.6
botocore==1.15.6
cattrs==1.0.0
docutils==0.15.2
jmespath==0.9.5
jsii==0.22.0
publication==0.0.3
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
elasticache_auto_discovery
pymemcache